+++ GraphLabs.Site.Models/Infrastructure/IFilterableByLabName.cs
@@ -1,0 +1,8 @@
+namespace GraphLabs.Site.Models.Infrastructure
+{
+    public interface IFilterableByLabName<TListModel, TModel>
+        where TListModel : IListModel<TModel>
+    {
+        TListModel FilterByLabName(string labname);
+    }
+}
