+++ GraphLabs.Site.Models/Infrastructure/IFilterableByUser.cs
@@ -1,0 +1,8 @@
+namespace GraphLabs.Site.Models.Infrastructure
+{
+    public interface IFilterableByUser<TListModel, TModel>
+        where TListModel : IListModel<TModel>
+    {
+        TListModel FilterByUser(string user);
+    }
+}
