+++ GraphLabs.Site/Views/Home/_NewsPartial.cshtml
@@ -3,7 +3,6 @@
 
 @if (!Model.Any())
 {
-    
     <p>Тут будут новости</p>
 }
 else
