--- GraphLabs.Site/Views/Schedule/Delete.cshtml
@@ -1,6 +1,0 @@
-﻿@{
-    ViewBag.Title = "Delete";
-}
-
-<!--<h2>Delete</h2>
-<button onclick="deleteTestPoolEntry(@item.Id);" type="button" id="delete_@item.Id">Удалить</button>-->
